+++ src/minishell.c
@@ -5,25 +5,52 @@
 /*                                                    +:+ +:+         +:+     */
 /*   By: rude-jes <rude-jes@student.42lausanne.c    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2024/03/26 17:06:10 by rude-jes          #+#    #+#             */
-/*   Updated: 2024/03/28 19:37:14 by rude-jes         ###   ########.fr       */
+/*   Created: 2024/03/26 19:56:01 by rude-jes          #+#    #+#             */
+/*   Updated: 2024/03/30 01:12:47 by rude-jes         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
 
-int	main(void)
+static char	*mini_gethostname(void)
+{
+	const char	hostname[] = "/etc/hostname";
+	char		*output;
+	char		*tmp;
+	char		*line;
+	int			fd;
+
+	fd = open(hostname, O_RDONLY);
+	if (fd < 0)
+		return (0);
+	line = ft_get_next_line(fd);
+	output = ft_strdup("");
+	while (line)
+	{
+		if (output)
+			tmp = output;
+		output = ft_strjoin(output, line);
+		if (!output)
+			return (0);
+		if (tmp)
+			free (tmp);
+		free(line);
+		line = ft_get_next_line(fd);
+	}
+	output[ft_strlen(output) - 1] = 0;
+	return (output);
+}
+
+t_minishell	*init_minishell(int argc, char **argv, char **envp)
 {
 	t_minishell	*minishell;
-	char		*input;
 
-	minishell = init_minishell();
-	while (true)
-	{
-		input = prompt(minishell);
-		if (!input)
-			crash_exit();
-		free(input);
-	}
-	exit (0);
+	minishell = galloc(sizeof(t_minishell));
+	if (!minishell)
+		crash_exit();
+	minishell->hostname = mini_gethostname();
+	minishell->argc = argc;
+	minishell->argv = argv;
+	minishell->envp = envp;
+	return (minishell);
 }
