+++ src/utils/exit_handler.h
@@ -1,20 +1,21 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   exit_handler.c                                     :+:      :+:    :+:   */
+/*   exit_handler.h                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: rude-jes <rude-jes@student.42lausanne.c    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2024/03/26 19:58:27 by rude-jes          #+#    #+#             */
-/*   Updated: 2024/03/30 01:10:35 by rude-jes         ###   ########.fr       */
+/*   Created: 2024/03/30 00:50:41 by rude-jes          #+#    #+#             */
+/*   Updated: 2024/03/30 01:15:23 by rude-jes         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "exit_handler.h"
+#ifndef EXIT_HANDLER_H
+# define EXIT_HANDLER_H
 
-void	crash_exit(void)
-{
-	ft_putstr_fd("minishell: ", STDERR_FILENO);
-	ft_putendl_fd(strerror(errno), STDERR_FILENO);
-	exit(1);
-}
+# include "../../libs/betterft/betterft.h"
+
+//	crash_exit: exit the program when unexpected error
+void	crash_exit(void);
+
+#endif
